requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
uuid
python-multipart
python-dotenv
datetime